New 3D rendering (j3d) & 2D text file data reader
Orson charts scatterplot has been replaced by j3d rendering. I also
figured out how to properly connect points for any 3D shape I want by
using actual (the default) j3d rendering. I made the program read in
from a text file to make the Temple. In the text file, it can be written
as an “arc” (which is a 3d cylinder), a “cone” (a 3d cone), a “sphe” (a
3d sphere), a “rect” (which is a 3d rectangle), and almost any 3D shape
the user wants. The text file can also take in an x point, y point,
width (or radius for cylinder/sphere), and a height (the z value from
the GUI). It can also take in lines (written as “line” in the text
file), which will have an x1, y1, x2, & y2. The (x1, y1) will be
connected to the (x2,y2) automatically, then it will automatically be
connected to the other lines to form a floor, roof, or any other
customized 3D shape you want. I also created the Mouse rotation,
zooming, and translating (pushing any direction). The Mouse rotation
still has some complicated glitch with the floor not rotated correctly
with all the 3D shapes, which we will try to fix in the next iteration.
I had to configure and initialize some TransformGroups for Transform3D
(to position/rotate everything). I have a separate TransformGroup and
Transform3D for the floor. Everything else (the shapes) is in the same
group. I needed to separate them to fix a strange glitch where
everything was invisible if I put the floor and the shapes in the same
group.